+++ resource-enumerate/documentation/js/__config/local-api-config.js
@@ -1,0 +1,3 @@
+export default (apiNs, {hostname}) => {
+  return `//${hostname}/apis/dev/v1.0/resource-enumerate.json`;
+};