+++ resource-enumerate/documentation/js/init-polyfill-service/index.js
@@ -1,0 +1,4 @@
+import {Polyfill} from '@aofl/polyfill-service';
+import polyfills from '../__config/polyfills';
+
+Polyfill.loadAll(polyfills);