+++ resource-enumerate/documentation/routes/home-element/examples/basic/template.js
@@ -1,0 +1,5 @@
+(context, html) => html`
+<p>Environment: ${context.environment}</p>
+<p>Response:</p>
+<pre>${JSON.stringify(context.response, null, 2)}</pre>
+`;