+++ resource-enumerate/index.js
@@ -1,5 +1,1 @@
-import resourceEnumerate from './src/ResourceEnumerate';
-
-export {
-  resourceEnumerate
-};
+export {default as ResourceEnumerate} from './src/resource-enumerate';